factors: Split HttpHandlerExecutor into variants
